requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
